+++ src/tests/EntityGraphQL.Tests/QueryTests/InputTypeTests.cs
@@ -153,6 +153,40 @@ public class InputTypeTests
     }
 
     [Fact]
+    public void SupportsQueryTypeAsInputType()
+    {
+        var schema = SchemaBuilder.Create<TestDataContext>();
+
+        schema.AddType<User>("User");
+        schema.AddType<Task>("Task");
+
+        schema.UpdateType<User>(type =>
+        {
+            type.AddField("id", x => x.Id, null);
+            type.AddField("tasks", x => x.Tasks, null);
+        });
+
+        schema.UpdateType<Task>(type =>
+        {
+            type.AddField("id", x => x.Id, null);
+            type.AddField("user", null).Resolve<TestDataContext>((p, ctx) => ctx.Users.FirstOrDefault(u => u.Id == p.Project.CreatedBy));
+        });
+
+        var taskInput = schema.AddInputType<Task>("TaskInput");
+        taskInput.AddField("id", x => x.Id, null);
+
+        var userInput = schema.AddInputType<User>("UserInput");
+        userInput.AddField("id", x => x.Id, null);
+        userInput.AddField("tasks", x => x.Tasks, null);
+
+        schema.Query().AddField("users", (ctx) => ctx.Users, "Get a users");
+        schema.Query().AddField("tasks", (ctx) => ctx.Tasks, "Get a tasks");
+
+        var result = schema.ToGraphQLSchemaString();
+        Assert.Contains("input UserInput {\n\tid: Int!\n\ttasks: [TaskInput!]\n}", result);
+    }
+
+    [Fact]
     public void SupportsQueryTypeAsInputTypeUsingReturns()
     {
         var schema = SchemaBuilder.Create<TestDataContext>();